fix(mcp): decode proto-JSON string-encoded int64 in list_backends
`list_backends` failed against every proto-first daemon with
"cannot unmarshal string into Go struct field Backend.uptimeSeconds of
type int64". When /v1/backends became the proto-first ListBackends RPC
(#354), grpc-gateway's protojson started serializing int64 as a QUOTED
STRING ("345600"), but the MCP client's Backend struct kept plain int64
fields (uptimeSeconds, gpus[].vramBytes) — so the whole response failed
to decode and the tool returned an error instead of the backend list.
This blocks every backend-topology operation (e.g. finding a GPU peer to
validate).

- Add flexInt64: an int64 that decodes from either a JSON number OR a
  quoted string, so it's robust across mixed daemon versions.
- Use it for Backend.UptimeSeconds and BackendGPU.VRAMBytes; cast to
  int64 at the two format call sites.
- Update the testdata fixture to the real proto-first wire (string-
  encoded int64) and fix the stale "handler is hand-coded, emits numbers"
  comment — that's exactly the assumption that broke.
- Tests: decode the string form (regression), the number form (mixed
  fleet), and null/empty → 0.

+++ internal/mcp/tools.go
@@ -1647,7 +1647,7 @@ func writeBackendDetail(b *strings.Builder, bk *Backend) {
 	}
 	fmt.Fprintf(b, "   Containers: %d running\n", bk.ContainerCount)
 	if bk.UptimeSeconds > 0 {
-		fmt.Fprintf(b, "   Uptime:     %s\n", formatUptime(bk.UptimeSeconds))
+		fmt.Fprintf(b, "   Uptime:     %s\n", formatUptime(int64(bk.UptimeSeconds)))
 	}
 	if bk.LastSeenAt != "" && bk.Type != "local" {
 		fmt.Fprintf(b, "   Last seen:  %s\n", bk.LastSeenAt)
@@ -1657,7 +1657,7 @@ func writeBackendDetail(b *strings.Builder, bk *Backend) {
 		for _, g := range bk.GPUs {
 			vram := ""
 			if g.VRAMBytes > 0 {
-				vram = fmt.Sprintf(" — %s VRAM", humanBytes(g.VRAMBytes))
+				vram = fmt.Sprintf(" — %s VRAM", humanBytes(int64(g.VRAMBytes)))
 			}
 			fmt.Fprintf(b, "     - %s %s%s\n", g.Vendor, g.ModelName, vram)
 		}
